requirements.txt
altgraph==0.17.5
customtkinter==5.2.2
darkdetect==0.8.0
ImageHash==4.3.2
numpy==2.4.1
packaging==26.0
pillow==12.1.0
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
PyWavelets==1.9.0
scipy==1.17.0
setuptools==80.10.1
tqdm==4.67.1
